+++ cmd/frontend/internal/app/BUILD.bazel
@@ -29,12 +29,12 @@ go_library(
         "//cmd/frontend/internal/app/otlpadapter",
         "//cmd/frontend/internal/app/router",
         "//cmd/frontend/internal/app/ui",
+        "//cmd/frontend/internal/auth/accessrequest",
         "//cmd/frontend/internal/routevar",
         "//cmd/frontend/oneclickexport",
         "//internal/actor",
         "//internal/api",
         "//internal/auth",
-        "//internal/auth/accessrequest",
         "//internal/auth/userpasswd",
         "//internal/authz",
         "//internal/cloneurls",
