+++ cmd/frontend/internal/auth/accessrequest/BUILD.bazel
@@ -1,11 +1,11 @@
-load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("//dev:go_defs.bzl", "go_test")
 
 go_library(
     name = "accessrequest",
     srcs = ["handlers.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/auth/accessrequest",
-    visibility = ["//:__subpackages__"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/auth/accessrequest",
+    visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//internal/actor",
         "//internal/auth/userpasswd",
@@ -22,13 +22,9 @@ go_library(
 
 go_test(
     name = "accessrequest_test",
-    timeout = "short",
     srcs = ["handlers_test.go"],
     embed = [":accessrequest"],
-    tags = [
-        # Test requires localhost database
-        "requires-network",
-    ],
+    tags = ["requires-network"],
     deps = [
         "//internal/conf",
         "//internal/database",
